refactor: user 도메인 리펙토링
로그인 유무에서 사용자 인증(auth)에 따른 Controller 분리
